funding: rename from fmgr, change references
